Move transaction impls to their type definitions file
