+++ substrate/frame/revive/src/evm/api/rpc_types.rs
@@ -18,7 +18,6 @@
 use super::*;
 use codec::{Decode, Encode};
 use scale_info::TypeInfo;
-use sp_core::{H160, U256};
 
 /// Configuration specific to a dry-run execution.
 ///
@@ -213,41 +212,6 @@ impl From<BlockNumberOrTag> for BlockNumberOrTagOrHash {
 	}
 }
 
-impl From<TransactionSigned> for TransactionUnsigned {
-	fn from(tx: TransactionSigned) -> Self {
-		use TransactionSigned::*;
-		match tx {
-			Transaction7702Signed(tx) => tx.transaction_7702_unsigned.into(),
-			Transaction4844Signed(tx) => tx.transaction_4844_unsigned.into(),
-			Transaction1559Signed(tx) => tx.transaction_1559_unsigned.into(),
-			Transaction2930Signed(tx) => tx.transaction_2930_unsigned.into(),
-			TransactionLegacySigned(tx) => tx.transaction_legacy_unsigned.into(),
-		}
-	}
-}
-
-#[test]
-fn can_deserialize_input_or_data_field_from_generic_transaction() {
-	let cases = [
-		("with input", r#"{"input": "0x01"}"#),
-		("with data", r#"{"data": "0x01"}"#),
-		("with both", r#"{"data": "0x01", "input": "0x01"}"#),
-	];
-
-	for (name, json) in cases {
-		let tx = serde_json::from_str::<GenericTransaction>(json).unwrap();
-		assert_eq!(tx.input.to_vec(), vec![1u8], "{}", name);
-	}
-
-	let err = serde_json::from_str::<GenericTransaction>(r#"{"data": "0x02", "input": "0x01"}"#)
-		.unwrap_err();
-	assert!(
-		err.to_string().starts_with(
-		"Both \"data\" and \"input\" are set and not equal. Please use \"input\" to pass transaction call data"
-		)
-	);
-}
-
 #[test]
 fn test_block_number_or_tag_or_hash_deserialization() {
 	let val: BlockNumberOrTagOrHash = serde_json::from_str("\"latest\"").unwrap();
@@ -265,255 +229,4 @@ fn test_block_number_or_tag_or_hash_deserialization() {
 		let val: BlockNumberOrTagOrHash = serde_json::from_str(s).unwrap();
 		assert_eq!(val, BlockNumberOrTagOrHash::BlockHash(H256([0xaau8; 32])));
 	}
-}
-
-impl GenericTransaction {
-	/// Create a new [`GenericTransaction`] from a signed transaction.
-	pub fn from_signed(tx: TransactionSigned, base_gas_price: U256, from: Option<H160>) -> Self {
-		Self::from_unsigned(tx.into(), base_gas_price, from)
-	}
-
-	/// The gas price that is actually paid (including priority fee).
-	pub fn effective_gas_price(&self, base_gas_price: U256) -> Option<U256> {
-		let effective_gas_price = if let Some(prio_price) = self.max_priority_fee_per_gas {
-			let max_price = self.max_fee_per_gas?;
-			Some(max_price.min(base_gas_price.saturating_add(prio_price)))
-		} else {
-			self.gas_price
-		};
-
-		// we do not implement priority fee as it does not map to tip well
-		// hence the effective gas price cannot be higher than the base price
-		effective_gas_price.map(|e| e.min(base_gas_price))
-	}
-
-	/// Create a new [`GenericTransaction`] from a unsigned transaction.
-	pub fn from_unsigned(
-		tx: TransactionUnsigned,
-		base_gas_price: U256,
-		from: Option<H160>,
-	) -> Self {
-		use TransactionUnsigned::*;
-		let mut tx = match tx {
-			TransactionLegacyUnsigned(tx) => GenericTransaction {
-				from,
-				r#type: Some(tx.r#type.as_byte()),
-				chain_id: tx.chain_id,
-				input: tx.input.into(),
-				nonce: Some(tx.nonce),
-				value: Some(tx.value),
-				to: tx.to,
-				gas: Some(tx.gas),
-				gas_price: Some(tx.gas_price),
-				..Default::default()
-			},
-			Transaction4844Unsigned(tx) => GenericTransaction {
-				from,
-				r#type: Some(tx.r#type.as_byte()),
-				chain_id: Some(tx.chain_id),
-				input: tx.input.into(),
-				nonce: Some(tx.nonce),
-				value: Some(tx.value),
-				to: Some(tx.to),
-				gas: Some(tx.gas),
-				access_list: Some(tx.access_list),
-				blob_versioned_hashes: tx.blob_versioned_hashes,
-				max_fee_per_blob_gas: Some(tx.max_fee_per_blob_gas),
-				max_fee_per_gas: Some(tx.max_fee_per_gas),
-				max_priority_fee_per_gas: Some(tx.max_priority_fee_per_gas),
-				..Default::default()
-			},
-			Transaction1559Unsigned(tx) => GenericTransaction {
-				from,
-				r#type: Some(tx.r#type.as_byte()),
-				chain_id: Some(tx.chain_id),
-				input: tx.input.into(),
-				nonce: Some(tx.nonce),
-				value: Some(tx.value),
-				to: tx.to,
-				gas: Some(tx.gas),
-				access_list: Some(tx.access_list),
-				max_fee_per_gas: Some(tx.max_fee_per_gas),
-				max_priority_fee_per_gas: Some(tx.max_priority_fee_per_gas),
-				..Default::default()
-			},
-			Transaction2930Unsigned(tx) => GenericTransaction {
-				from,
-				r#type: Some(tx.r#type.as_byte()),
-				chain_id: Some(tx.chain_id),
-				input: tx.input.into(),
-				nonce: Some(tx.nonce),
-				value: Some(tx.value),
-				to: tx.to,
-				gas: Some(tx.gas),
-				gas_price: Some(tx.gas_price),
-				access_list: Some(tx.access_list),
-				..Default::default()
-			},
-			Transaction7702Unsigned(tx) => GenericTransaction {
-				from,
-				r#type: Some(tx.r#type.as_byte()),
-				chain_id: Some(tx.chain_id),
-				input: tx.input.into(),
-				nonce: Some(tx.nonce),
-				value: Some(tx.value),
-				to: Some(tx.to),
-				gas: Some(tx.gas),
-				access_list: Some(tx.access_list),
-				authorization_list: tx.authorization_list,
-				max_fee_per_gas: Some(tx.max_fee_per_gas),
-				max_priority_fee_per_gas: Some(tx.max_priority_fee_per_gas),
-				..Default::default()
-			},
-		};
-		tx.gas_price = tx.effective_gas_price(base_gas_price);
-		tx
-	}
-
-	/// Convert to a [`TransactionUnsigned`].
-	pub fn try_into_unsigned(self) -> Result<TransactionUnsigned, ()> {
-		match self.r#type.unwrap_or_default().0 {
-			TYPE_LEGACY => Ok(TransactionLegacyUnsigned {
-				r#type: TypeLegacy {},
-				chain_id: self.chain_id,
-				input: self.input.to_bytes(),
-				nonce: self.nonce.unwrap_or_default(),
-				value: self.value.unwrap_or_default(),
-				to: self.to,
-				gas: self.gas.unwrap_or_default(),
-				gas_price: self.gas_price.unwrap_or_default(),
-			}
-			.into()),
-			TYPE_EIP1559 => Ok(Transaction1559Unsigned {
-				r#type: TypeEip1559 {},
-				chain_id: self.chain_id.unwrap_or_default(),
-				input: self.input.to_bytes(),
-				nonce: self.nonce.unwrap_or_default(),
-				value: self.value.unwrap_or_default(),
-				to: self.to,
-				gas: self.gas.unwrap_or_default(),
-				gas_price: self.max_fee_per_gas.unwrap_or_default(),
-				access_list: self.access_list.unwrap_or_default(),
-				max_fee_per_gas: self.max_fee_per_gas.unwrap_or_default(),
-				max_priority_fee_per_gas: self.max_priority_fee_per_gas.unwrap_or_default(),
-			}
-			.into()),
-			TYPE_EIP2930 => Ok(Transaction2930Unsigned {
-				r#type: TypeEip2930 {},
-				chain_id: self.chain_id.unwrap_or_default(),
-				input: self.input.to_bytes(),
-				nonce: self.nonce.unwrap_or_default(),
-				value: self.value.unwrap_or_default(),
-				to: self.to,
-				gas: self.gas.unwrap_or_default(),
-				gas_price: self.gas_price.unwrap_or_default(),
-				access_list: self.access_list.unwrap_or_default(),
-			}
-			.into()),
-			TYPE_EIP4844 => Ok(Transaction4844Unsigned {
-				r#type: TypeEip4844 {},
-				chain_id: self.chain_id.unwrap_or_default(),
-				input: self.input.to_bytes(),
-				nonce: self.nonce.unwrap_or_default(),
-				value: self.value.unwrap_or_default(),
-				to: self.to.unwrap_or_default(),
-				gas: self.gas.unwrap_or_default(),
-				max_fee_per_gas: self.max_fee_per_gas.unwrap_or_default(),
-				max_fee_per_blob_gas: self.max_fee_per_blob_gas.unwrap_or_default(),
-				max_priority_fee_per_gas: self.max_priority_fee_per_gas.unwrap_or_default(),
-				access_list: self.access_list.unwrap_or_default(),
-				blob_versioned_hashes: self.blob_versioned_hashes,
-			}
-			.into()),
-			TYPE_EIP7702 => Ok(Transaction7702Unsigned {
-				r#type: TypeEip7702 {},
-				chain_id: self.chain_id.unwrap_or_default(),
-				input: self.input.to_bytes(),
-				nonce: self.nonce.unwrap_or_default(),
-				value: self.value.unwrap_or_default(),
-				to: self.to.unwrap_or_default(),
-				gas: self.gas.unwrap_or_default(),
-				max_fee_per_gas: self.max_fee_per_gas.unwrap_or_default(),
-				max_priority_fee_per_gas: self.max_priority_fee_per_gas.unwrap_or_default(),
-				access_list: self.access_list.unwrap_or_default(),
-				authorization_list: self.authorization_list,
-			}
-			.into()),
-			_ => Err(()),
-		}
-	}
-}
-
-#[test]
-fn from_unsigned_works_for_legacy() {
-	let base_gas_price = U256::from(10);
-	let tx = TransactionUnsigned::from(TransactionLegacyUnsigned {
-		chain_id: Some(U256::from(1)),
-		input: Bytes::from(vec![1u8]),
-		nonce: U256::from(1),
-		value: U256::from(1),
-		to: Some(H160::zero()),
-		gas: U256::from(1),
-		gas_price: U256::from(10),
-		..Default::default()
-	});
-
-	let generic = GenericTransaction::from_unsigned(tx.clone(), base_gas_price, None);
-	assert_eq!(generic.gas_price, Some(U256::from(10)));
-
-	let tx2 = generic.try_into_unsigned().unwrap();
-	assert_eq!(tx, tx2);
-}
-
-#[test]
-fn from_unsigned_works_for_1559() {
-	let base_gas_price = U256::from(10);
-	let tx = TransactionUnsigned::from(Transaction1559Unsigned {
-		chain_id: U256::from(1),
-		input: Bytes::from(vec![1u8]),
-		nonce: U256::from(1),
-		value: U256::from(1),
-		to: Some(H160::zero()),
-		gas: U256::from(1),
-		gas_price: U256::from(20),
-		max_fee_per_gas: U256::from(20),
-		max_priority_fee_per_gas: U256::from(1),
-		..Default::default()
-	});
-
-	let generic = GenericTransaction::from_unsigned(tx.clone(), base_gas_price, None);
-	assert_eq!(generic.gas_price, Some(U256::from(10)));
-
-	let tx2 = generic.try_into_unsigned().unwrap();
-	assert_eq!(tx, tx2);
-}
-
-#[test]
-fn from_unsigned_works_for_7702() {
-	let base_gas_price = U256::from(10);
-	let tx = TransactionUnsigned::from(Transaction7702Unsigned {
-		chain_id: U256::from(1),
-		input: Bytes::from(vec![1u8]),
-		nonce: U256::from(1),
-		value: U256::from(1),
-		to: H160::zero(),
-		gas: U256::from(1),
-		max_fee_per_gas: U256::from(20),
-		max_priority_fee_per_gas: U256::from(1),
-		authorization_list: vec![AuthorizationListEntry {
-			chain_id: U256::from(1),
-			address: H160::from_low_u64_be(42),
-			nonce: U256::from(0),
-			y_parity: U256::from(1),
-			r: U256::from(1),
-			s: U256::from(2),
-		}],
-		..Default::default()
-	});
-
-	let generic = GenericTransaction::from_unsigned(tx.clone(), base_gas_price, None);
-	assert_eq!(generic.gas_price, Some(U256::from(10)));
-
-	let tx2 = generic.try_into_unsigned().unwrap();
-	assert_eq!(tx, tx2);
 }
